refactor: make the results output formatting more specific
