+++ packages/code-analyzer-core/src/output-formats/results/json-output-format.ts
@@ -1,6 +1,6 @@
-import {CodeLocation, RunResults, Violation} from "../results";
-import {OutputFormatter, CODE_ANALYZER_CORE_NAME} from "../output-format";
-import {Rule, SeverityLevel} from "../rules";
+import {CodeLocation, RunResults, Violation} from "../../results";
+import {OutputFormatter, CODE_ANALYZER_CORE_NAME} from "../../output-format";
+import {Rule, SeverityLevel} from "../../rules";
 
 /**
  * Type associated with the JSON Output Format
@@ -86,9 +86,9 @@ export type JsonCodeLocationOutput = {
 }
 
 /**
- * Formatter for JSON Output Format
+ * Formatter for Results JSON Output Format
  */
-export class JsonOutputFormatter implements OutputFormatter {
+export class ResultsJsonOutputFormatter implements OutputFormatter {
     format(results: RunResults): string {
         const resultsOutput: JsonResultsOutput = toJsonResultsOutput(results);
         return JSON.stringify(resultsOutput, undefined, 2);
