+++ packages/code-analyzer-core/src/output-formats/results/xml-output-format.ts
@@ -1,12 +1,12 @@
-import {RunResults} from "../results";
+import {RunResults} from "../../results";
 import * as xmlbuilder from "xmlbuilder";
-import {OutputFormatter, CODE_ANALYZER_CORE_NAME} from "../output-format";
+import {OutputFormatter, CODE_ANALYZER_CORE_NAME} from "../../output-format";
 import {JsonResultsOutput, toJsonResultsOutput} from "./json-output-format";
 
 /**
- * Formatter for XML Output Format
+ * Formatter for Results XML Output Format
  */
-export class XmlOutputFormatter implements OutputFormatter {
+export class ResultsXmlOutputFormatter implements OutputFormatter {
     format(results: RunResults): string {
         // XML and JSON output formats are very similar, so leveraging the same data structure from JSON for now.
         const resultsOutput: JsonResultsOutput = toJsonResultsOutput(results);
@@ -75,5 +75,6 @@ export class XmlOutputFormatter implements OutputFormatter {
         }
 
         return violationsNode.end({ pretty: true, allowEmpty: true });
+
     }
 }
